backend/requirements.txt
fastapi
uvicorn
vosk
pydub
pydantic
requests
numpy
opencv-python-headless
python-multipart
python-dotenv
